pyproject.toml
[tool.poetry]
name = "mireapprove"
version = "0.1.0"
description = ""
authors = ["MireApprove Contributors"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.121.2"
uvicorn = "^0.38.0"
python-dotenv = "^1.2.1"
aiohttp = "^3.13.2"
cryptography = "^46.0.3"
beautifulsoup4 = "^4.14.2"
asyncpg = "^0.30.0"
email-validator = "^2.3.0"
pydantic = {extras = ["email"], version = "^2.12.5"}
requests = "^2.32.5"
python-dateutil = "^2.9.0"
apscheduler = "^3.11.1"
protobuf = "^6.33.1"
bbpb = "^1.4.2"
redis = "^5.0.0"
openpyxl = "^3.1.0"
pyotp = "^2.9.0"
pyzbar = "^0.1.9"
pillow = "^12.1.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.24.0"
pytest-cov = "^6.0.0"
httpx = "^0.28.0"
detect-secrets = "^1.5.0"
black = "^25.12.0"
isort = "^7.0.0"
flake8 = "^7.3.0"
autoflake = "^2.3.1"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.isort]
profile = "black"

[tool.black]
line-length = 88

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
